fix uid in Sonarr and Radarr import scripts
