+++ Scripts/Flow/Applications/Sonarr/Sonarr - Trigger Manual Import 2.0.js
@@ -1,25 +1,25 @@
-import { Radarr } from 'Shared/Radarr';
+import { Sonarr } from 'Shared/Sonarr';
 
 /**
- * @name Radarr - Trigger Manual Import 2.0
- * @uid cd8926f3-a9ab-4fa7-a36e-cceb8ad58987
- * @description Trigger Radarr to manually import the media, run after last file in folder moved.
+ * @name Sonarr - Trigger Manual Import 2.0
+ * @uid 8fe63a47-4a97-4b08-99f4-368d29279ad2
+ * @description Trigger Sonarr to manually import the media, run after last file in folder moved.
  * @author iBuSH
  * @revision 2
- * @param {string} URL Radarr root URL and port (e.g. http://radarr:7878)
- * @param {string} ApiKey Radarr API Key
+ * @param {string} URL Sonarr root URL and port (e.g. http://sonarr:8989)
+ * @param {string} ApiKey Sonarr API Key
  * @param {string} ImportPath The output path for import triggering (default Working File)
  * @param {bool} UseUnmappedPath Whether to Unmap the path to the original FileFlows Server path when using nodes in different platforms (e.g. Docker and Windows)
  * @param {bool} MoveMode Import mode 'copy' or 'move' (default copy)
  * @output Command sent
  */
 function Script(URL, ApiKey, ImportPath, UseUnmappedPath, MoveMode) {
-  URL = URL || Variables['Radarr.Url'] || Variables['Radarr.URI'];
+  URL = URL || Variables['Sonarr.Url'] || Variables['Sonarr.URI'];
   URL = URL.replace(/\/+$/g, "");
-  ApiKey = ApiKey || Variables['Radarr.ApiKey'];
+  ApiKey = ApiKey || Variables['Sonarr.ApiKey'];
   ImportPath = ImportPath || Variables.file.FullName;
   const ImportMode = MoveMode ? 'move' : 'copy';
-  const radarr = new Radarr(URL, ApiKey);
+  const sonarr = new Sonarr(URL, ApiKey);
 
   if (UseUnmappedPath) ImportPath = Flow.UnMapPath(ImportPath);
 
@@ -27,16 +27,17 @@ function Script(URL, ApiKey, ImportPath, UseUnmappedPath, MoveMode) {
   Logger.ILog(`Triggering Path: ${ImportPath}`);
   Logger.ILog(`Import Mode: ${ImportMode}`);
 
+
   let commandBody = {
       path: ImportPath,
       importMode: ImportMode
     };
 
-  let response = radarr.sendCommand('downloadedMoviesScan', commandBody)
+  let response = sonarr.sendCommand('downloadedEpisodesScan', commandBody)
 
   // Wait for the completion of the scan
   let commandId = response['id']
-  let importCompleted = radarr.waitForCompletion(commandId);
+  let importCompleted = sonarr.waitForCompletion(commandId);
   if (!importCompleted) {
       Logger.WLog('Import failed');
       return -1;
